backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
motor==3.3.2
pymongo==4.6.1
redis==5.0.1
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
prometheus-client==0.19.0
python-json-logger==2.0.7
httpx==0.25.1
pytest-mock==3.12.0
email-validator==2.1.0